Crate meenle_noonle

Meenle_Noonle is my software renderer demo, built to help me learn things I didn’t previously know. It is a small Rust library with no dependencies, targeting wasm. It exports functions to act on an internal frame buffer (render, draw_line), as well as a function to get a pointer to the frame buffer itself (get_buffer).

Modules

demo

Demonstration of the meenle-noonle library. To use, first call generate_background, then pick your mesh with set_mesh, and use get_buffer to get the frame buffer where the output will be rendered. Call render_spin every frame to update the frame buffer. NOTE: The demo is not thread safe. You must wait for set_mesh to finish execution before rendering.

meshes

Some sample meshes for demonstration.

Structs

Mat3x3

Simple 3x3 Matrix for graphics maths.

Mesh

Mesh of triangles.

Pixel

Pixel for the frame buffer. RGBA color, to match HTML canvas’ buffer format.

Tri

Triangle.

Vec3

Enums

Axis

Constants

DEPTH_RANGE 🔒

HEIGHT

MODEL_Z_RANGE 🔒

WIDTH

Statics

BG_BUFFER 🔒

Frame buffer with the pretty background pattern. Used to clear the scene.

BUFFER 🔒

Frame buffer.

Functions

draw_line

Uses Bresenham’s algorithm to draw a line.

fill_buffer

Fills the frame buffer with a pretty pattern.

generate_background

Generates the pretty background pattern.

get_buffer

Gets a pointer to the frame buffer.

plot_pixel 🔒

Plots a single pixel into the frame buffer.

render

Renders the demo into the frame buffer.

Type Definitions

Vertex